+++ src/main/components/form/claim-details-validator.ts
